Remove backend when not in current well-known docs (fix #52)
In general, a backend can be removed when it was not referenced in any of the well-known documents that were crawled in the current run.

However, what happens if a well-known service is temporarily unavailable? The backends that used to be referenced from it would be deleted too. To avoid this, all failed attempts of downloading a well-known document are logged. The backends that used to belong to one of the "failed services" are then exempt from deletion.

But what happens if a well-known service is permanently unavailable? Will the backend data be left in the database forever? No, the unsuccessfulCrawls mechanism (which is executed directly afterwards) will take care of them eventually.

Because it may be a bit confusing at first:
DELETE all that are [not in allIndividualBackends AND not in allFailedServices]
is the same as
KEEP all that are [in allIndividualBackends OR in allFailedServices]