Update openeo dependencies (js-commons, vue-comp.)
The FeatureList component has been moved from the js-commons to the vue-components repo, hence the change in the DiscoverSection component.

Note that, as of time of this commit, vue-components has not been officially released as 1.0.0-beta.3 yet (the latest release is beta.2). But the changes are already committed in the `dev` branch of that repo (which I run locally via `npm link`) and will be part of the next release. Therefore I already put it in the package.json anyway.