Use map from vue-components for spatial extent (close #54)
Instead of drawing its own, the Hub can rely on the work in the vue-components repo.

Changes to previous appearance/behaviour:
- map is now only 300px high (instead of 400px) -- probably better anyway to get more stuff into view at the same time.
- bounding box coordinates are not printed (extent is *only* shown in map) -- I kind of miss it, but probably most people won't.
- map may zoom closer if the extent is small -- definitely an improvement.

scrollWheelZoom=false and noWrap=true have been retained by passing the appropriate `mapOptions` (wrapAroundAntimeridian is already false by default but in case that changes I explicitly pass it anyway).

Although one may think so after this change, the `vue2-leaflet` dependency can't be removed (or changed to just `leaflet`) because the BboxChooser (which I plan to use again soon) also uses it.