Integrate v1.0's /file_formats endpoint (part of #53)
Into crawling (crawl.js), raw data processing (first two hunks in dbqueries.js) and auto-suggest listing (last hunk in dbqueries.js).

The endpoint can simply be added to the list of endpoint candidates because the crawl script will only try to access those endpoints that are listed as supported in the capabilities document (which is downloaded first).

The `outputFormats` field is now handled in a way so that it basically means `allFileFormats`: For v0.4 backends it continues to contain just the output formats (since there are no input formats); for v1.0 backends it now contains the combined input and output formats.

The UI in Backend.vue is changed accordingly, too.