Adapt web editor URL (closes #58, related to #53)
There are now two deployments of the web editor, one that supports v0.4.x and one for v1.0.x, so the Hub needs to forward to the correct one depending on the `api_version` field.

For now, I left the v0.3-related behaviour (e.g. webEditorUnavailable) in there because it may be handy again in the future and doesn't really hurt at the moment.