Only start server when DB con. has been established
Otherwise it can happen that restify already serves the frontend, but kills itself a few seconds later because it couldn't establish a database connection -- and then the user already has the app available and tries to make requests, which of course all fail, which isn't nice. And the server shouldn't serve anything before there's a db connection anyway.