Rename outputFormats to fileFormats.output (part of #53)
In commit 695eb89 I made the existing `outputFormats` field behave like an `allFileFormats` field, i.e. it contained the combined input AND output formats. That (1) didn't make much sense semantically -- the field should've been renamed, too. But (2) secondly and more importantly: Input and output formats should be displayed (and searchable) independently of each other.

Part (1) is fixed completely in this commit. For part (2) this commit sets the basis (corrected database structure) and realises the independent display and search for output formats. The realisation for input formats will follow in the next commit.

I decided to rewrite the GET_ALL_BACKENDS pipeline so that the file formats are always saved in the new API v1.0 format, i.e. the old field `outputFormats` is renamed to `fileFormats` and contains two sub-objects with the keys `input` and `output`. This makes the GET_ALL_{IN|OUT}PUT_FORMATS_WITH_COUNT pipelines a lot easier because they don't have to deal with two separate structures -- you can see that in dbqueries.js three lines (124-126) could be merged into one. (And in the long run, if we drop support for v0.4, only one pipeline needs to be changed instead of three.)

This renaming and restructuring incurred subsequent changes in Backend.vue and DiscoverSection.vue.