Implement handling of fileFormats.input (part of #53)
This is the second part of what was begun in the previous commit 4d413e9. See also the commit message there.

Essentially, everything that existed for `fileFormats.output` was duplicated with `out` replaced by `in` -- the behaviour is entirely analogous.

One exception: In the dbqueries.js, the second item of the `$ifNull` is the empty object because input formats didn't exist in the 0.4.x versions of the openEO API spec.