Display status of new production flag (part of #53)
According to the openEO API v1.0.0-rc.2 the flag defaults to `true`, hence the `$ifNull` handling in the dbqueries.

But in the frontend, the affirmation that it is indeed production-ready should not be displayed for v0.4.x backends because the flag didn't exist back then, hence the extra check in the second `v-if`.

-> for v0.4.x backends, simply nothing is shown.

I hope that the dark yellow is good to read on all displays (it's fine on mine).