Show well-known URL instead of direct URLs (fix #69)
The direct URLs shouldn't be shown (at least not so prominently), in favour of the well-known URLs. Those are the same though for every backend of the same group, so it's enough to show them once per group.

This fix doesn't look overly pretty, but it's okay enough for now. A long-term solution will be discussed in #70.

I committed this change directly to master because it fixes a "high priority" issue.