Fix jumping headings when filtering
The first item had a special CSS so that it didn't feature the extra top margin all the others had. But when it was filtered out (but only `display:none`d, not removed from the DOM), the now "visibly first" item *did* have that margin, which caused some unpleasant jumping around of the headings. Now that's better and the overall padding/margin distribution was improved too (give *all* items that top margin so the extra bottom margin for the previous text becomes unnecessary; plus add more space when things are uncollapsed).