Handle changed imported endpoint format (fix #77)
The EndpointChooser receives the endpoint list via a prop, which is populated by the DiscoverSection, with the content of FeatureList.features, which is imported from openeo-vue-components. That data structure used to contain one array per group, but that was changed to objects in Open-EO/openeo-vue-components@594dd73. So when I updated the dependency in 91721ae, this mechanism broke. Now it works again.