Correct grouping of coll./proc. in crawl script (fix #78)
Of course, these must be grouped by the full unique key of the collections, not only by the `id` field (which is just the _name_ of the collection/process within its backend, not an actual unique ID across ALL backends). Otherwise, only one of e.g. the many `filter_bbox` processes survives this stage.