Apply filtering per backend (rather than per provider)
Move the `checkFilters` method from DiscoverSection into BackendGroup so that the filtering can be done there, with DiscoverSection's `filters` passed via a prop.

During the move, what was essentially changed in the code was removing the initial `backends.some` from every filter condition, and doing a little hack with an IIFE for the endpoints, that's it. (To be precise, `backends.some(b => ` was replaced with `(` so that the parantheses remain.)

The hiding of non-matching backends is done via the `enabled` property of the `Tab`s (for this, openeo-vue-components v1.0.0-rc.3 is required). Very handy.

To fully hide providers of which absolutely no backend matches, the `doesAnyFilterMatch` method is new; it simply checks the filters for all the backends. This means a bit of double computation, but I don't think that's too bad given the clean code of this solution. The `v-show` now sits on the inner <div> of BackendGroup, no longer on the <li> of DiscoverSection, which isn't the nicest, but it is by far the most simple solution so again, I don't think that's too bad.