Move process graph validation to server (#46)
And thus make it work properly! When the filter-by-process-graph feature was introduced in dbdd0cc it didn't work right just yet because there's not enough data on the client side. Thus now the move to the server.

On the downside, because the code that determines whether to filter out a backend is decentralised (located directly in the BackendGroup component rather than one layer higher), we now have to do 1 ajax request per BackendGroup, i.e. pasting a process graph results in (currently) 8 calls being made, each of which transfers mostly the same data. Not nice, but the simplest solution for now.

This resulted in a new endpoint for the Hub's API. I tried to implement it compliant to `/validation` of the openEO API spec. It's a bit hacky, because the backend URL to which the validation request shall belong must be transferred too and I (ab)used the `links` field for that. But I think it's bearable. As long as only 1 backend URL is submitted in this way, the reply IS compliant to the openEO API spec. When more than 1 backend URL is submitted, the reply will be an array of the documents that would've been returned had each validation been submitted independently. (Note that an error in ANY of the individual validations may crash the entire request though.)

It's on purpose that the new endpoint is NOT listed in the capabilities endpoint or the Readme, as it's not intended for public use (yet).