Display input and output file formats in same component
The FileFormats component has badges to indicate which ones are input and which ones are output, so it's nicer to have them together in one component.
The defaults for showInput and showOutput are both `true` anyway, but stating that explicitly makes it clearer and also more robust in case this is changed again in the future.