Remove isSearchResult prop from Backend component
That prop was set to true when the Backend component was used in the context of the Search section -- which has been removed from the Hub ages ago. So it's about time to get rid of that Überbleibsel.

The code in the `created` hook can be deleted too since sorting can be achieved with the Collections and Processes components themselves (which use the SearchableList component under the hood).