Use collapsing mechanism of imported components
Works for all except SupportedFeatures (but that one isn't collapsed by default anyway and there isn't really a need to) and Billing (here I left in my own implementation and made sure to copy the CSS of vue-components).

The definition list syntax is gone now, but that's okay.

Some CSS had to be adjusted to make it still look the same as before.