Upgrade mongodb driver (which fixes #91)
I checked the migration guide [1] against the code and did a test run of all discovery functionality the Hub currently offers and didn't find any issues, so the upgrade should be fine. As a byproduct, this also fixes the issue of #91, which is essentially the same issue as #90, which apparently wasn't fixed by 36bb389 as intended.

[1] https://github.com/mongodb/node-mongodb-native/blob/HEAD/etc/notes/CHANGES_4.0.0.md